fixed backend type mismatch

+++ cinema_backend/src/main/java/com/group27/cinema_backend/controller/CheckoutController.java
@@ -178,7 +178,7 @@ public class CheckoutController {
         resp.setSuccess(true);
         resp.setMessage("Booking completed successfully");
         resp.setBookingId(booking.getId());
-        resp.setTotalPrice(booking.getTotalPrice());
+        resp.setTotalPrice(booking.getTotalPrice().doubleValue());
         resp.setBookedSeatIds(request.getSeatIds());
 
         return ResponseEntity.ok(resp);
